Add squidlib integration to make use of all its cool stuff SOme minor refactors

+++ core/src/com/PaulChew/Triangle.java
@@ -140,7 +140,7 @@ public class Triangle extends ArraySet<Pnt> implements Comparable<Triangle> {
 
     @Override
     public int hashCode () {
-        return (int)(idNumber^(idNumber>>>32));
+        return idNumber ^ ( idNumber >>> 32);
     }
 
     @Override
